requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
apscheduler==3.10.4
email-validator==2.1.0
bcrypt==4.1.2
python-dotenv==1.0.1
geopy==2.4.1 